apps/py-api/pyproject.toml
[project]
name = "py-api"
version = "1.0.0"
description = "FastAPI backend for the NSP app (Phase 7 / Chapter 15)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "ortools>=9.15",
    "nsp-core",
    "sqlmodel>=0.0.22",
    "aiosqlite>=0.20",
    "sse-starlette>=2.1",
    "prometheus-client>=0.21",
    "pyyaml>=6.0",
    "greenlet>=3.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "ruff>=0.7",
    "mypy>=1.13",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/py_api"]

[tool.uv.sources]
nsp-core = { path = "../py-cp-sat/nsp-core", editable = true }

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM", "TID"]
# B008: Depends(...) as default-argument is the standard FastAPI pattern.
ignore = ["E501", "B008"]

[tool.mypy]
python_version = "3.12"
strict = true

[[tool.mypy.overrides]]
module = ["ortools.*", "sse_starlette.*", "nsp_core.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"
asyncio_mode = "auto"
